Eliminate a few more memory leaks
git-svn-id: http://openmx.psyc.virginia.edu/svn/trunk@3205 df83fd30-4cca-4c36-ab9d-e8b5583ccbd1

+++ src/omxExportBackendState.cpp
@@ -60,7 +60,7 @@ void omxExportResults(omxState *currentState, MxRList *out)
 		    }
 		}
 
-		if(OMX_DEBUG) { mxLog("Final Calculation of Algebra %d Complete.", index); }
+		if(OMX_DEBUG) { mxLog("Final Calculation of Algebra %d Complete.", (int) index); }
 		SET_VECTOR_ELT(algebras, index, algebra);
 	}
 	if(OMX_DEBUG) { mxLog("All Algebras complete."); }
